refactor: ExtruderControlPanel.vue
