+++ src/store/printer/getters.ts
@@ -772,28 +772,6 @@ export const getters: GetterTree<PrinterState, RootState> = {
         return output
     },
 
-    getToolchangeMacros: (state, getters) => {
-        const macros = getters['getMacros']
-        const tools: PrinterStateToolchangeMacro[] = []
-
-        macros
-            .filter((macro: any) => macro.name.toUpperCase().match(/^T\d+/))
-            .forEach((macro: any) =>
-                tools.push({
-                    name: macro.name,
-                    active: macro.variables.active ?? false,
-                    color: macro.variables.color ?? macro.variables.colour ?? null,
-                })
-            )
-
-        return tools.sort((a, b) => {
-            const numberA = parseInt(a.name.slice(1))
-            const numberB = parseInt(b.name.slice(1))
-
-            return numberA - numberB
-        })
-    },
-
     getKinematics: (state) => {
         if (!state.configfile?.settings?.printer) return false
